requirements.txt
python-telegram-bot
pillow
requests
# for .env file support
python-dotenv
